+++ src/cache/src/Annotation/Cacheable.php
@@ -11,6 +11,7 @@ declare(strict_types=1);
  */
 namespace Hyperf\Cache\Annotation;
 
+use Attribute;
 use Hyperf\Cache\CacheListenerCollector;
 use Hyperf\Di\Annotation\AbstractAnnotation;
 use Hyperf\Di\Annotation\AnnotationCollector;
@@ -19,6 +20,7 @@ use Hyperf\Di\Annotation\AnnotationCollector;
  * @Annotation
  * @Target({"METHOD"})
  */
+#[Attribute(Attribute::TARGET_METHOD)]
 class Cacheable extends AbstractAnnotation
 {
     /**
@@ -57,9 +59,9 @@ class Cacheable extends AbstractAnnotation
      */
     public $collect = false;
 
-    public function __construct($value = null)
+    public function __construct(...$value)
     {
-        parent::__construct($value);
+        parent::__construct(...$value);
         if ($this->ttl !== null) {
             $this->ttl = (int) $this->ttl;
         }