+++ src/circuit-breaker/src/Exception/TimeoutException.php
@@ -13,7 +13,7 @@ namespace Hyperf\CircuitBreaker\Exception;
 
 class TimeoutException extends CircuitBreakerException
 {
-    public function __construct(string $message = '', $result)
+    public function __construct(string $message, $result)
     {
         parent::__construct($message);
         $this->result = $result;