+++ src/grpc/src/StatusCode.php
@@ -21,12 +21,12 @@ class StatusCode
     /**
      * The operation completed successfully.
      */
-    const OK = 0;
+    public const OK = 0;
 
     /**
      * The operation was cancelled (typically by the caller).
      */
-    const CANCELLED = 1;
+    public const CANCELLED = 1;
 
     /**
      * Unknown error.  An example of where this error may be returned is
@@ -35,7 +35,7 @@ class StatusCode
      * errors raised by APIs that do not return enough error information
      * may be converted to this error.
      */
-    const UNKNOWN = 2;
+    public const UNKNOWN = 2;
 
     /**
      * Client specified an invalid argument.  Note that this differs
@@ -43,7 +43,7 @@ class StatusCode
      * that are problematic regardless of the state of the system
      * (e.g., a malformed file name).
      */
-    const INVALID_ARGUMENT = 3;
+    public const INVALID_ARGUMENT = 3;
 
     /**
      * Deadline expired before operation could complete.  For operations
@@ -52,17 +52,17 @@ class StatusCode
      * successful response from a server could have been delayed long
      * enough for the deadline to expire.
      */
-    const DEADLINE_EXCEEDED = 4;
+    public const DEADLINE_EXCEEDED = 4;
 
     /**
      * Some requested entity (e.g., file or directory) was not found.
      */
-    const NOT_FOUND = 5;
+    public const NOT_FOUND = 5;
 
     /**
      * Some entity that we attempted to create (e.g., file or directory) already exists.
      */
-    const ALREADY_EXISTS = 6;
+    public const ALREADY_EXISTS = 6;
 
     /**
      * The caller does not have permission to execute the specified
@@ -72,13 +72,13 @@ class StatusCode
      * used if the caller cannot be identified (use UNAUTHENTICATED
      * instead for those errors).
      */
-    const PERMISSION_DENIED = 7;
+    public const PERMISSION_DENIED = 7;
 
     /**
      * Some resource has been exhausted, perhaps a per-user quota, or
      * perhaps the entire file system is out of space.
      */
-    const RESOURCE_EXHAUSTED = 8;
+    public const RESOURCE_EXHAUSTED = 8;
 
     /**
      * Operation was rejected because the system is not in a state
@@ -97,7 +97,7 @@ class StatusCode
      * should be returned since the client should not retry unless
      * they have first fixed up the directory by deleting files from it.
      */
-    const FAILED_PRECONDITION = 9;
+    public const FAILED_PRECONDITION = 9;
 
     /**
      * The operation was aborted, typically due to a concurrency issue
@@ -106,7 +106,7 @@ class StatusCode
      * <p>See litmus test above for deciding between FAILED_PRECONDITION,
      * ABORTED, and UNAVAILABLE.
      */
-    const ABORTED = 10;
+    public const ABORTED = 10;
 
     /**
      * Operation was attempted past the valid range.  E.g., seeking or
@@ -124,19 +124,19 @@ class StatusCode
      * so that callers who are iterating through
      * a space can easily look for an OUT_OF_RANGE error to detect when they are done.
      */
-    const OUT_OF_RANGE = 11;
+    public const OUT_OF_RANGE = 11;
 
     /**
      * Operation is not implemented or not supported/enabled in this service.
      */
-    const UNIMPLEMENTED = 12;
+    public const UNIMPLEMENTED = 12;
 
     /**
      * Internal errors.  Means some invariants expected by underlying
      * system has been broken.  If you see one of these errors,
      * something is very broken.
      */
-    const INTERNAL = 13;
+    public const INTERNAL = 13;
 
     /**
      * The service is currently unavailable.  This is a most likely a
@@ -147,23 +147,23 @@ class StatusCode
      * <p>See litmus test above for deciding between FAILED_PRECONDITION,
      * ABORTED, and UNAVAILABLE.
      */
-    const UNAVAILABLE = 14;
+    public const UNAVAILABLE = 14;
 
     /**
      * Unrecoverable data loss or corruption.
      */
-    const DATA_LOSS = 15;
+    public const DATA_LOSS = 15;
 
     /**
      * The request does not have valid authentication credentials for the
      * operation.
      */
-    const UNAUTHENTICATED = 16;
+    public const UNAUTHENTICATED = 16;
 
     /**
      * @see https://grpc.github.io/grpc/core/md_doc_statuscodes.html
      */
-    const HTTP_CODE_MAPPING = [
+    public const HTTP_CODE_MAPPING = [
         self::OK => 200,
         self::CANCELLED => 499,
         self::UNKNOWN => 500,