+++ src/ide-helper/src/Metadata.php
@@ -1,0 +1,24 @@
+<?php
+
+declare(strict_types=1);
+/**
+ * This file is part of Hyperf.
+ *
+ * @link     https://www.hyperf.io
+ * @document https://hyperf.wiki
+ * @contact  group@hyperf.io
+ * @license  https://github.com/hyperf/hyperf/blob/master/LICENSE
+ */
+namespace Hyperf\IDEHelper;
+
+use Doctrine\Common\Annotations\Reader;
+use ReflectionClass;
+
+class Metadata
+{
+    public function __construct(
+        public Reader $reader,
+        public ReflectionClass $reflection
+    ) {
+    }
+}